refactor: move link recover and duplicate remove upgrader to migration job
