+++ packages/nocodb/src/db/BaseModelSqlv2.ts
@@ -10230,9 +10230,13 @@ export function getCompositePkValue(primaryKeys: Column[], row) {
   if (typeof row !== 'object') return row;
 
   if (primaryKeys.length > 1) {
-    return primaryKeys.map((c) =>
-      (row[c.title] ?? row[c.column_name])?.toString?.().replaceAll('_', '\\_'),
-    ).join('___');
+    return primaryKeys
+      .map((c) =>
+        (row[c.title] ?? row[c.column_name])
+          ?.toString?.()
+          .replaceAll('_', '\\_'),
+      )
+      .join('___');
   }
 
   return (
