+++ packages/nocodb/src/services/columns.service.ts
@@ -47,7 +47,6 @@ import {
   createHmAndBtColumn,
   createOOColumn,
   generateFkName,
-  randomID,
   sanitizeColumnName,
   validateLookupPayload,
   validatePayload,
