--- packages/nocodb/src/version-upgrader/upgraders/0227000_dupColMergeUpgrader.ts
@@ -1,6 +1,0 @@
-// this migration is only applicable in ee
-import type { NcUpgraderCtx } from '~/version-upgrader/NcUpgrader';
-
-export default async (_: NcUpgraderCtx) => {
-  // do nothing
-};
